# Conflicts:
#	src/main/java/seedu/duke/parser/Parser.java

+++ src/main/java/seedu/duke/command/RemoveShelfCommand.java
@@ -3,14 +3,12 @@ package seedu.duke.command;
 import seedu.duke.command.exception.ShelfNotExistException;
 import seedu.duke.model.ShelfList;
 
-import java.util.regex.Pattern;
-
 public class RemoveShelfCommand extends Command {
     public static final String REMOVE_DATA_ARGS_FORMAT_STRING = "remove shlv/SHELF_NAME";
     public static final String REMOVE_STRING = "remove";
     public static final String PARSE_REMOVE_SUCCESS_MESSAGE_FORMAT = "shelfname: %s\n";
     private static final String REMOVE_COMPLETE_MESSAGE =
-            "This shelf %s has been deleted.";
+            "Shelf \"%s\" has been deleted.";
     private final String shelfName;
 
     public RemoveShelfCommand(String shelfName) {
@@ -20,6 +18,10 @@ public class RemoveShelfCommand extends Command {
     @Override
     public String execute() throws ShelfNotExistException {
         try {
+            if (ShelfList.getShelfList().existShelf(shelfName)) {
+                throw new IllegalArgumentException("Cannot remove shelf with existing items");
+            }
+            
             ShelfList.getShelfList().deleteShelf(shelfName);
             return String.format(REMOVE_COMPLETE_MESSAGE, shelfName);
         } catch (seedu.duke.model.exception.ShelfNotExistException e) {
